+++ apps/api/src/app/endpoints/public/public.controller.ts
@@ -129,8 +129,7 @@ export class PublicController {
               type,
               unitPrice,
               value,
-              valueInBaseCurrency,
-              SymbolProfile: assetProfile
+              valueInBaseCurrency
             };
           }
         );