+++ apps/api/src/interceptors/transform-data-source-in-response/transform-data-source-in-response.interceptor.ts
@@ -80,10 +80,6 @@ export class TransformDataSourceInResponseInterceptor<
             paths: [
               'activities[*].assetProfile.dataSource',
               'activities[*].dataSource',
-
-              /* @deprecated */
-              'activities[*].SymbolProfile.dataSource',
-
               'assetProfile.dataSource',
               'benchmarks[*].dataSource',
               'errors[*].dataSource',
@@ -92,10 +88,6 @@ export class TransformDataSourceInResponseInterceptor<
               'holdings[*].assetProfile.dataSource',
               'holdings[*].dataSource',
               'items[*].dataSource',
-
-              /* @deprecated */
-              'SymbolProfile.dataSource',
-
               'watchlist[*].dataSource'
             ]
           });